feat(x1): change the ratelimit class to new file
